Move to src layout; remove scripts
