+++ src/repology/repology_cli.py
@@ -24,7 +24,7 @@ from bs4 import BeautifulSoup
 import numpy as np
 import pandas as pd
 from tabulate import tabulate
-from sbomnix.utils import (
+from common.utils import (
     LOG,
     LOG_SPAM,
     set_log_verbosity,
