+++ src/vulnxscan/__init__.py
@@ -1,0 +1,3 @@
+# SPDX-FileCopyrightText: 2022 Technology Innovation Institute (TII)
+#
+# SPDX-License-Identifier: Apache-2.0
