+++ tests/compare_sboms.py
@@ -15,7 +15,7 @@ import sys
 import pathlib
 import json
 import pandas as pd
-from sbomnix.utils import LOG, df_to_csv_file, set_log_verbosity
+from common.utils import LOG, df_to_csv_file, set_log_verbosity
 
 ###############################################################################
